refactor(instance): split instance_state/mod.rs's test module into tests.rs
CI's file_size_invariant caught it: the new creator-ACL tests pushed
mod.rs to 904 LOC (max 750). Moves the existing #[cfg(test)] mod tests
body out to instance_state/tests.rs (test files are exempt from the
LOC guard, same pattern as lifecycle/tests.rs) — no test or behavior
changes, pure relocation.

Agend-Agent: gapfix-dev
Agend-Task: t-20260702132521582830-56872-2
Agend-Branch: feat/delete-instance-creator-acl
Agend-Issued-At: 2026-07-02T15:07:40.393252+00:00
